Changed database package structure to match the offline branch
