+++ libs/pandautils/src/main/java/com/instructure/pandautils/room/appdatabase/AppDatabase.kt
@@ -1,10 +1,12 @@
-package com.instructure.pandautils.room
+package com.instructure.pandautils.room.appdatabase
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import androidx.room.TypeConverters
+import com.instructure.pandautils.room.Converters
+import com.instructure.pandautils.room.appdatabase.daos.*
+import com.instructure.pandautils.room.appdatabase.entities.*
 import com.instructure.pandautils.room.daos.*
-import com.instructure.pandautils.room.entities.*
 
 @Database(
     entities = [
