+++ libs/pandautils/src/main/java/com/instructure/pandautils/room/appdatabase/daos/MediaCommentDao.kt
@@ -1,10 +1,10 @@
-package com.instructure.pandautils.room.daos
+package com.instructure.pandautils.room.appdatabase.daos
 
 import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Insert
 import androidx.room.Update
-import com.instructure.pandautils.room.entities.MediaCommentEntity
+import com.instructure.pandautils.room.appdatabase.entities.MediaCommentEntity
 
 @Dao
 interface MediaCommentDao {
