+++ libs/pandautils/src/main/java/com/instructure/pandautils/room/appdatabase/daos/PendingSubmissionCommentDao.kt
@@ -1,8 +1,8 @@
-package com.instructure.pandautils.room.daos
+package com.instructure.pandautils.room.appdatabase.daos
 
 import androidx.room.*
-import com.instructure.pandautils.room.entities.PendingSubmissionCommentEntity
-import com.instructure.pandautils.room.model.PendingSubmissionCommentWithFileUploadInput
+import com.instructure.pandautils.room.appdatabase.entities.PendingSubmissionCommentEntity
+import com.instructure.pandautils.room.appdatabase.model.PendingSubmissionCommentWithFileUploadInput
 
 @Dao
 interface PendingSubmissionCommentDao {
