+++ libs/pandautils/src/main/java/com/instructure/pandautils/room/appdatabase/entities/AuthorEntity.kt
@@ -1,4 +1,4 @@
-package com.instructure.pandautils.room.entities
+package com.instructure.pandautils.room.appdatabase.entities
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
