+++ libs/pandautils/src/main/java/com/instructure/pandautils/room/appdatabase/entities/FileUploadInputEntity.kt
@@ -1,8 +1,8 @@
-package com.instructure.pandautils.room.entities
+package com.instructure.pandautils.room.appdatabase.entities
 
 import androidx.room.Entity
 import androidx.room.PrimaryKey
-import java.util.Random
+import java.util.*
 
 @Entity
 data class FileUploadInputEntity(
