+++ libs/pandautils/src/main/java/com/instructure/pandautils/room/appdatabase/model/PendingSubmissionCommentWithFileUploadInput.kt
@@ -1,12 +1,12 @@
-package com.instructure.pandautils.room.model
+package com.instructure.pandautils.room.appdatabase.model
 
 import androidx.room.Embedded
 import androidx.room.Relation
 import com.instructure.canvasapi2.models.postmodels.CommentSendStatus
 import com.instructure.canvasapi2.models.postmodels.FileUploadWorkerData
 import com.instructure.canvasapi2.models.postmodels.PendingSubmissionComment
-import com.instructure.pandautils.room.entities.FileUploadInputEntity
-import com.instructure.pandautils.room.entities.PendingSubmissionCommentEntity
+import com.instructure.pandautils.room.appdatabase.entities.FileUploadInputEntity
+import com.instructure.pandautils.room.appdatabase.entities.PendingSubmissionCommentEntity
 import java.util.*
 
 data class PendingSubmissionCommentWithFileUploadInput(
